refactor: move MCP route wiring out of handlers.go into mcp/router
Other apiserver modules (llm-gateway, a2a, email-relay, ...) keep
their route registration inside their own subpackage and let the
global handlers.go call a single InitRoutes/RegisterRoutes function.
MCP was the odd one out: ~80 lines of bootstrapping plus two helper
functions (initMCPRoutes / mountMCPRoutes) lived directly inside
handlers.go.

This commit aligns MCP with the rest of the codebase:
- New apiserver/pkg/mcp/router package owns InitRoutes (composes the
  server with unified + builtin tools and instructions) and
  MountRoutes (binds SSE / streamable HTTP / health / index onto a
  Gin engine under the configured base path).
- handlers.go shrinks to a single mcprouter.InitRoutes(engine) call,
  matching the llm-gateway / a2a pattern.
- handlers_mcp_test.go moves to apiserver/pkg/mcp/router/router_test.go
  with no behavioural change; the only edit is renaming the call from
  the lowercase mountMCPRoutes to the now-exported MountRoutes.

No functional change. All existing tests still pass (router pkg
coverage 63.6%).

Made-with: Cursor
